Revert "Bug/23 (#25)"
This reverts commit cc32a1247b68cfcff073da17f9b209a321f609aa.
