+++ .gitignore
@@ -1,3 +1,2 @@
 # User files
 *.user
-.idea
